Fix slf4j api and add missing btm scope test
slf4j was coming from btm.  So fixed it to use matching api to simple
impl.
btm was listed in test section but was missing scope which meant it was
compile and transcient if anyone uses mybatis-cdi so fixed that to test
scope so the test dependency is no longer picked up by mistake.