Delayed filtering of Mappers to instantiate to a moment where all Mappers have been found.
On TomEE it is possible that a ProcessInjectionTarget is processed prior to a processAnnotatedType has found the actual Mapper.
In those cases the Mapper would not be added to the this.mappers Set.
This would result in an UnsatisfiedResolutionException when the CDI Container wants to instantiate the InjectionTarget.

This change delayes checking the ProcessInjectionTargets to the afterBeanDiscovery-method, when all Mappers are found.

fixes #40