[tests] Re-enable tests after splitting test execution
Due to behaviour of injections, split so these run successfully.  It's not normal one would have both types setup and it appears that the setup is getting mixed.  While this still might point to some underlying issue, it resolves so we run all tests and the source is no different than it originally was.