[ci] Update readme and suggested support levels
drop jdk 7 as mybatis doesn't support it and we moved to require jdk 8 long time ago.
Mark not supported more clearly.  For example we do not support 1.0 and won't.  Moved preferred to 2.0 (pre jakarta EE changes).  Add generic jdk columns showing 8, 11, and 17.  As jdk 17 is not passing any tests, mark all as N for supported until we can get it working...